synq: extract LSP semantic-tokens delta encoder from host.rs
The ~100-line semantic-tokens delta encoder is pure LSP-protocol
wire-format code (UTF-16 code-unit counting, delta-line encoding,
legend index emission) and belongs in its own file rather than in
the LspHost facade.

- new lsp/semantic_tokens_codec.rs holds encode_semantic_tokens and
  its utf16_len helper
- host.rs drops the encoding block entirely and calls the extracted
  function directly; one fewer reason to pull TokenCategory / DocLen
  into host.rs

No behavior change. 244 unit + 34 doctests + 25 integration tests
pass on --features lsp.
